Refactor deserialization integration tests to `tests/` using `from_str`, avoiding private functions. Nicely organize `sets` feature tests in `tests/lib.rs`
