+++ tests/lib.rs
@@ -1,5 +1,11 @@
 #![recursion_limit = "512"]
 
+pub mod deserialize;
 pub mod emit;
 pub mod parse;
 pub mod ser;
+
+#[cfg(feature = "sets")]
+pub mod deserialize_sets;
+#[cfg(feature = "sets")]
+pub mod parse_sets;
